+++ docker/scripts/build_docker.sh
@@ -13,10 +13,10 @@ fi
 # exit on errors
 trap 'exit' ERR
 
-# need env variable GITHUB_PAT set in /root/.bash_profile of host for r.sif to build.
+# need env variable GITHUB_PAT set in /root/.bashrc of host for r.sif to build.
 # this is a personal access token from github with read:packages scope and will have
 # to be updated every 30 days.
-if [ $1 == "r" ]; then
+if [ $1 == "r.sif" ]; then
     source ~/.bashrc
     if [ -z "$GITHUB_PAT" ]; then
         echo "GITHUB_PAT not set"